Feat: Populates Database with ICD-11 comorbidities on docker startup
